chore: move infos components in file
